# swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DepressoSpeech – Unified API Documentation",
    "description": "Comprehensive API documentation for the DepressoSpeech Depression Severity Prediction System.\n\n## Services\n\n| Service | Port | Description |\n|---------|------|-------------|\n| **MindScope Backend** | 8000 | Core application API – authentication, assessments, file management, doctor/admin dashboards |\n| **DepressoSpeech ML Model** | 8001 | ML inference API – PHQ-8 prediction from audio using eGeMAPS + MFCC + SBERT features |\n| **Frontend (Vite)** | 5173 | React SPA – patient/doctor/admin user interface |\n| **Swagger Docs** | 8080 | This documentation page |\n\n## Authentication\n- **Backend API**: JWT Bearer tokens via `/api/v1/auth/login`\n- **ML Model API**: Optional API key via `X-API-Key` header\n- **Admin**: Separate admin login via `/api/v1/auth/admin/login`\n\n## Architecture\nThe system follows a microservices pattern:\n1. Frontend sends requests to Backend (proxied via Vite in dev)\n2. Backend orchestrates business logic, stores data in SQLite\n3. Backend delegates ML inference to the Model service\n4. Model service processes audio and returns PHQ-8 predictions",
    "version": "2.0.0",
    "contact": {
      "name": "DepressoSpeech Team – KMIT",
      "email": "koushik@kmit.in"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
...
    },
    {
...
    }
  ],
...
  "paths": {
...
    "/predict": {
      "post": {
        "tags": ["ML Prediction"],
        "summary": "Predict PHQ-8 from audio",
        "description": "Upload a single audio file and receive a PHQ-8 depression severity prediction. Audio is preprocessed (VAD, chunking), features extracted (eGeMAPS + MFCC + text), and the model predicts a score (0-24). Rate limited to 30 requests/minute.",
        "operationId": "predictSingle",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "name": "participant_id", "in": "query", "schema": { "type": "string", "default": "unknown" }, "description": "Optional participant identifier (1-64 chars)" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary", "description": "Audio file (wav, mp3, flac, ogg, m4a, webm)" }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PHQ-8 prediction result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PredictionResponse" }
              }
            }
          },
          "413": { "description": "File too large (max 100 MB)" },
          "422": { "description": "Invalid file type or content" },
          "429": { "description": "Rate limit exceeded" },
          "503": { "description": "Model not loaded" }
        }
      }
    },
    "/predict/extended": {
      "post": {
        "tags": ["ML Prediction"],
        "summary": "Extended prediction with confidence",
        "description": "Extended prediction returning PHQ-8 score plus confidence intervals (MC Dropout), audio quality metrics, and behavioral features. Rate limited to 20 requests/minute.",
        "operationId": "predictExtended",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          { "name": "participant_id", "in": "query", "schema": { "type": "string", "default": "unknown" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Extended prediction result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExtendedPredictionResponse" }
              }
            }
          }
        }
      }
    },
    "/predict/batch": {
      "post": {
        "tags": ["ML Prediction"],
        "summary": "Batch prediction",
        "description": "Upload multiple audio files (max 20) and receive PHQ-8 predictions for each. Failed files are reported but don't block others. Rate limited to 10 requests/minute.",
        "operationId": "predictBatch",
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": { "type": "string", "format": "binary" },
                    "description": "List of audio files (max 20)"
                  }
                },
                "required": ["files"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch prediction results",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BatchPredictionResponse" }
              }
            }
          },
          "422": { "description": "Too many files" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from `/api/v1/auth/login`"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Optional API key for ML model (set via DEPRESSO_API_KEY env var)"
      }
    },
    "schemas": {
...
      "PredictionResponse": {
        "type": "object",
        "properties": {
          "participant_id": { "type": "string" },
          "phq8_score": { "type": "number", "minimum": 0, "maximum": 24, "description": "Predicted PHQ-8 score" },
          "severity": { "type": "string", "enum": ["Minimal", "Mild", "Moderate", "Moderately Severe", "Severe"] },
          "num_chunks": { "type": "integer", "description": "Number of audio chunks processed" },
          "timestamp": { "type": "string", "format": "date-time" }
        }
      },
      "ExtendedPredictionResponse": {
        "type": "object",
        "properties": {
          "participant_id": { "type": "string" },
          "phq8_score": { "type": "number", "minimum": 0, "maximum": 24 },
          "severity": { "type": "string" },
          "num_chunks": { "type": "integer" },
          "inference_time_s": { "type": "number" },
          "timestamp": { "type": "string", "format": "date-time" },
          "confidence": {
            "type": "object",
            "properties": {
              "mean": { "type": "number" },
              "std": { "type": "number" },
              "ci_lower": { "type": "number" },
              "ci_upper": { "type": "number" }
            }
          },
          "audio_quality": {
            "type": "object",
            "properties": {
              "rms": { "type": "number" },
              "snr_db": { "type": "number" },
              "speech_prob": { "type": "number" },
              "quality": { "type": "string" }
            }
          },
          "behavioral": {
            "type": "object",
            "description": "Behavioral features extracted from audio"
          }
        }
      },
      "BatchPredictionResponse": {
        "type": "object",
        "properties": {
          "predictions": { "type": "array", "items": { "$ref": "#/components/schemas/PredictionResponse" } },
          "total": { "type": "integer" },
          "failed": { "type": "array", "items": { "type": "string" } }
        }
      },
...
    }
  }
}